+++ src/features/NovelView/NovelViewPresentation.tsx
@@ -12,28 +12,8 @@ interface NovelViewPresentationProps {
   childrenPanel: Sentence[];
   startIndexParent: number;
   setStartIndexParent: React.Dispatch<React.SetStateAction<number>>;
-  evaluationGoodCountParent: number;
-  setEvaluationGoodCountParent: React.Dispatch<React.SetStateAction<number>>;
-  evaluationStayCountParent: number;
-  setEvaluationStayCountParent: React.Dispatch<React.SetStateAction<number>>;
-  isGoodEvaluatedParent: boolean;
-  isStayEvaluatedParent: boolean;
   startIndexChildren: number;
   setStartIndexChildren: React.Dispatch<React.SetStateAction<number>>;
-  evaluationGoodCountChildren: number;
-  setEvaluationGoodCountChildren: React.Dispatch<React.SetStateAction<number>>;
-  commentCountChildren: number;
-  setCommentCountChildren: React.Dispatch<React.SetStateAction<number>>;
-  evaluationStayCountChildren: number;
-  setEvaluationStayCountChildren: React.Dispatch<React.SetStateAction<number>>;
-  isGoodEvaluatedChildren: boolean;
-  isStayEvaluatedChildren: boolean;
-  evaluationGoodCountMain: number;
-  setEvaluationGoodCountMain: React.Dispatch<React.SetStateAction<number>>;
-  evaluationStayCountMain: number;
-  setEvaluationStayCountMain: React.Dispatch<React.SetStateAction<number>>;
-  isGoodEvaluatedMain: boolean;
-  isStayEvaluatedMain: boolean;
   hasMainPanelEvaluation: boolean;
   textCount: number;
   onPost: (newSentence: string) => Promise<void>;
